requirements.txt
requests_kerberos
python-dotenv
flask
json2xml
xmltodict
python-dotenv